minor improvements (#11)
Seems like German was already well translated. Not sure if you want to use the word Ward or its translation= "Stadtviertel", "Stadtteil" or "Statbezirk".
"clear" was translated as the adjective "Klar" but here it means to clear or delete, I thought the more adequate translation is "Löschen"
Good job for all the rest!